locks

+++ scripts/17-git/install.sh
@@ -1,6 +1,11 @@
 #!/bin/bash
 
-echo " - GIT..."
+echo " - Git..."
 CONFS=$(dirname $0)/confs
 
+LOCK=/tmp/workstation.lock.git
+[ -f ${LOCK} ] && return
+
 apt-get install -y git git-flow > /dev/null 2>&1
+
+> ${LOCK}
